restore old objects.js for diff hygiene
